+++ Kernel/Modules/AgentTicketQueue.pm
@@ -491,7 +491,6 @@ sub BuildQueueView {
         UserID        => $Self->{UserID},
         QueueID       => $Self->{QueueID},
         ShownQueueIDs => $Param{QueueIDs},
-        Filter        => $Param{Filter},
     );
 
     # build output ...
@@ -512,6 +511,8 @@ sub _MaskQueueView {
     my $QueueIDOfMaxAge = $Param{QueueIDOfMaxAge} || -1;
     my %AllQueues       = %{ $Param{AllQueues} };
     my %Counter;
+    my %Totals;
+    my $HaveTotals = 0;    # flag for "Total" in index backend
     my %UsedQueue;
     my @ListedQueues;
     my $Level       = 0;
@@ -530,6 +531,7 @@ sub _MaskQueueView {
         push @ListedQueues, $QueueRef;
         my %Queue = %$QueueRef;
         my @Queue = split /::/, $Queue{Queue};
+        $HaveTotals ||= exists $Queue{Total};
 
         # remember counted/used queues
         $UsedQueue{ $Queue{Queue} } = 1;
@@ -543,14 +545,26 @@ sub _MaskQueueView {
             else {
                 $QueueName .= '::' . $Queue[$_];
             }
-            if ( !$Counter{$QueueName} ) {
-                $Counter{$QueueName} = 0;
+            if ( !exists $Counter{$QueueName} ) {
+                $Counter{$QueueName} = 0;    # init
+                $Totals{$QueueName}  = 0;
             }
-            $Counter{$QueueName} = $Counter{$QueueName} + $Queue{Count};
-            if ( $Counter{$QueueName} && !$Queue{$QueueName} && !$UsedQueue{$QueueName} ) {
+            my $Total = $Queue{Total} || 0;
+            $Counter{$QueueName} += $Queue{Count};
+            $Totals{$QueueName}  += $Total;
+            if (
+                ( $Counter{$QueueName} || $Totals{$QueueName} )
+                && !$Queue{$QueueName}
+                && !$UsedQueue{$QueueName}
+                )
+            {
+                # IMHO, this is purely pathological--TicketAcceleratorIndex
+                # sorts queues by name, so we should never stumble across one
+                # that we have not seen before!
                 my %Hash = ();
                 $Hash{Queue} = $QueueName;
                 $Hash{Count} = $Counter{$QueueName};
+                $Hash{Total} = $Total;
                 for ( sort keys %AllQueues ) {
                     if ( $AllQueues{$_} eq $QueueName ) {
                         $Hash{QueueID} = $_;
@@ -571,7 +585,8 @@ sub _MaskQueueView {
         # replace name of CustomQueue
         if ( $Queue{Queue} eq 'CustomQueue' ) {
             $Counter{$CustomQueue} = $Counter{ $Queue{Queue} };
-            $Queue{Queue} = $CustomQueue;
+            $Totals{$CustomQueue}  = $Totals{ $Queue{Queue} };
+            $Queue{Queue}          = $CustomQueue;
         }
         my @QueueName = split /::/, $Queue{Queue};
         my $ShortQueueName = $QueueName[-1];
@@ -628,8 +643,16 @@ sub _MaskQueueView {
         }
 
         # QueueStrg
-        $QueueStrg .= $Self->{LayoutObject}->Ascii2Html( Text => $ShortQueueName )
-            . " ($Counter{$Queue{Queue}})";
+        $QueueStrg .= $Self->{LayoutObject}->Ascii2Html( Text => $ShortQueueName );
+
+        # If the index backend supports totals, we show total tickets
+        # as well as unlocked ones in the form  "QueueName (total / unlocked)"
+        if ( $HaveTotals && ( $Totals{ $Queue{Queue} } != $Counter{ $Queue{Queue} } ) ) {
+            $QueueStrg .= " ($Totals{$Queue{Queue}}/$Counter{$Queue{Queue}})";
+        }
+        else {
+            $QueueStrg .= " ($Counter{$Queue{Queue}})";
+        }
 
         $QueueStrg .= '</a></li>';